+++ t/1140-isf-public-schedule-report-metadata-audit.t
@@ -54,6 +54,7 @@ use FSM::Support::ISFPublicInterfaceContract qw(
     isf_public_interface_schedule_report_top_level_keys
     isf_public_interface_schedule_report_transaction_port_binding_keys
     isf_public_interface_schedule_report_transaction_port_binding_site_kind_values
+    isf_public_interface_schedule_report_transaction_wait_keys
     isf_public_interface_schedule_report_transaction_stage_keys
     isf_public_interface_schedule_report_transaction_stage_kind_values
     isf_public_interface_schedule_report_transaction_keys
@@ -144,6 +145,10 @@ sub assert_schedule_report_metadata {
         [
             schedule_report_transaction_port_binding_keys =>
                 isf_public_interface_schedule_report_transaction_port_binding_keys(),
+        ],
+        [
+            schedule_report_transaction_wait_keys =>
+                isf_public_interface_schedule_report_transaction_wait_keys(),
         ],
         [
             schedule_report_fanin_group_required_keys =>